refactor(vue-media): Extract timestamp utilities and fix code review issues
- Create shared timestamp.ts utility with formatting functions
- Fix token staleness in HLS.js by accessing authStore.token directly
- Add error handling for initMediaSession failures
- Add aria-labels for accessibility on form elements
- Remove duplicate toApiTimestamp and formatTimestamp functions
- Remove unused CSS classes
